feat: add latest news to default rsd homepage
