+++ frontend/components/home/imperial/index.tsx
@@ -1,5 +1,5 @@
-// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
-// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2023 - 2024 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 - 2024 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -15,7 +15,7 @@ import MainContentImperialCollege from './MainContentImperialCollege'
 const pageTitle = `Home | ${app.title}`
 const pageDesc = 'The Research Software Directory is designed to show the impact research software has on research and society. We stimulate the reuse of research software and encourage proper citation of research software to ensure researchers and RSEs get credit for their work.'
 
-export default function ImperialCollegeHome({counts}: HomeProps) {
+export default function ImperialCollegeHome({counts,news}: HomeProps) {
   return (
     <>
       <PageMeta
@@ -26,7 +26,7 @@ export default function ImperialCollegeHome({counts}: HomeProps) {
 
       <section className="flex-1 flex flex-col text-secondary-content  bg-[url('/images/campus_south_ken.jpg')] bg-contain bg-no-repeat bg-center bg-black bg-scroll">
         <AppHeader />
-        <MainContentImperialCollege counts={counts} />
+        <MainContentImperialCollege counts={counts} news={news}/>
         <AppFooter/>
       </section>
     </>
