+++ frontend/components/home/rsd/HomepageDivider.tsx
@@ -1,0 +1,10 @@
+// SPDX-FileCopyrightText: 2024 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2024 Netherlands eScience Center
+//
+// SPDX-License-Identifier: Apache-2.0
+
+export default function HomepageDivider() {
+  return (
+    <div className="w-full max-w-screen-xl mx-auto border-t border-[#90909060] mt-[80px]"></div>
+  )
+}
