+++ frontend/components/home/rsd/config.ts
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: 2022 Dusan Mijatovic (dv4all)
 // SPDX-FileCopyrightText: 2022 dv4all
-// SPDX-FileCopyrightText: 2023 Dusan Mijatovic (Netherlands eScience Center)
-// SPDX-FileCopyrightText: 2023 Netherlands eScience Center
+// SPDX-FileCopyrightText: 2023 - 2024 Dusan Mijatovic (Netherlands eScience Center)
+// SPDX-FileCopyrightText: 2023 - 2024 Netherlands eScience Center
 //
 // SPDX-License-Identifier: Apache-2.0
 
@@ -54,6 +54,11 @@ export const config = {
       url: 'mailto:rsd@esciencecenter.nl?subject=Question about RSD',
       target: '_blank'
     },
+    news:{
+      label: 'More news',
+      url: '/news',
+      target: '_self'
+    }
   }
 }
